Distinguish between "price" and "value" when explaining *price* volatility
Similar to how we would not say cars are volatile if large price
swings for cars happen often, we shouldn't say Bitcoin is volatile
because its price is volatile.
It's important to distinguish between the underlying technology
that is Bitcoin, and the price of the tradable currency unit
within this network. The currency units may fluctuate often and
greatly in price, but the Bitcoin technology itself is not volatile
because of this.

Additionally, I've used the lowercase "bitcoin" when talking about
the currency *unit* rather than the "Bitcoin" technology/network.