Updated Bitcoin <title> for Google. Minor punct.
I added a <head> to index.html along with a <title> so Bitcoin is clearly described on Google. I chose the title "Bitcoin - Open-source P2P Digital Currency" because it describes Bitcoin clearly and concisely and that's what the definition of Bitcoin is on the frontpage. In other words, people will know exactly what Bitcoin is when it shows up on Google.

I also made minor grammatical and punctuation fixes: A colon before the list of features, a hyphen in "open-source", two hyphens in "peer-to-peer" and I changed "and a software" to "and software".