Remove press representatives
Regardless of the dislaimers on this page, anyone listed here are going
to be perceived as representatives of Bitcoin. This has already caused
un needed political discussions on bitcointalk and elsewhere.

I know journalists want this, but lets force them to do a bit of work
themselves. Such as emailing the press list.

That way bitcoin.org remains unpolitical and uncontroversial.