Remove Mike Hearn from press List for Promoting Radical Positions
Mike Hearn has promoted radical ideas possible detrimental to Bitcoin:

http://www.youtube.com/watch?v=mD4L7xDNCmA&feature=player_embedded#t=1005s

At 16:45 he presents a slide with the title "Banking but no banks"

He said:  
"Can we build a society that looks like the society we have today but is entirely free of banks?  There is nothing wrong with banks.  (applause)  People like that idea.  You know the people who work at banks they are usually good people and trying hard in many cases but they have privleged positions and are often abused do things we do not agree with.  In general it would be better if we didn't need them."