QA: Simplify Comparison Between Local And Remote Site Files
This commit adds an extra step to the build process to SHA256 checksum
all the files that are built purely based on repository contents.  The
checksums file is uploaded to the webserver with the rest of the site
content.

A separate target is added to the Makefile to compare the remote
sha256sums file to a locally-built file to see if they differ.  This
allows us to detect when a remote build may have gone astray.