Refactor how content is presented in base.html
A better way to present content is to add a container to the content
itself instead of having a separate content.html file with one generic
class.

So a file like about-us.html can have a wrapping class of
“content_about_us” because it decouples the css and making the code
more module.

A better way is to create a generic class with all the styles in less.
Then add this class like this.

.site_generic() {
  a { styles };
  etc…

.content_about_us {
  .site_generic;
}

Second note:
For the time being I remove breadcrumbs.