Add builder system
The problem is that having global css rules make it hard for anyone to
easily change things on a page. Specially to experiment with new
designs.

So you end up with a pile of code that nobody wants to touch. When I
started this redesign, the decision to delete all the css code and also
for some extend refactoring all html code for the pages was an easy
decision. It would maybe not take longer but would be so much more
boring and painful that it was easy to delete.

**Solution**
The solution is to have a file that continue all the global rules and
as few of them as possible. This file is “site.less”.

site.less contains the ‘canvas’, that is the rules for the area you can
design in.

Then inside the same folder you have the “/builder” folder. This folder
contains all the building code for the most comment elements on the
page.

So to create a new page you do this:
1. Create a new file in _templates folder.
2. Wrap this file with a div.
3. Add class to this wrapping div, class=‘content_standard’
4. Start writing. Lookup the markup for the different elements when you
need to.

If you are a designer and want to design your own page:
1. Create a new file in _templates folder.
2. Create a new file in _less/content/content_filename.less
4. Start with a plank slate. No global rules exist. So start creating.
5. If you want to add elements to the page, just use the builders.

this will create the css:
.builder__content__title(yourcontinerclass);

This is the markup:
    <h1 class="content_yourcontinerclass--title">{% translate pagetitle
%}</h1>
    <h2 class="content_yourcontinerclass--subtitle">{% translate
summary %}</h2>

**Results**
This way of designing is very good for open source-projects because now
anyone can start creating a pages without the need to know global rules
because non exist.

And if they want to use pre-made elements they just need to use
‘content_standard’ or more custom way of adding the builders one by one.

This system also make it easier to delete legacy code and mountain a
growing code base.

**Negative**
This system is need to use Gzip. The system produce a lot of repeating
code that takes up memory. This is a perfect problem for Gzip that was
designed for this kinds of tasks. But without Gzip the stylesheet can
become very big in size.