Add disclaimer to bitcoin-core index page
There is some confusion in the community that Bitcoin Core and bitcoin.org are the same project
and thus share the same policy and views. This disclaimer is to make it clear that Bitcoin Core is
a separate project and just occupy /bitcoin-core on the website for the main purpose of hosting
downloads.