Changes from Currency concept to Commodity concept.
This change is due to the rulling of Florida (vs) Espinoza (Case # F14-2923)
and the Article 78 against New York Division of Financial Services (case to be heard 3/30/2017 in New York City)

For the good of (B|b)itcoin, it is important that the documentation reflect the true nature of what it is.
Currency is defined by the fact that a central authority creates and destroy a currency.
A commodity exist regardles of what the central authority decrete; so bitcoin is really the first "Virtual Commodity" ever created.

The documentation need to reflect the fact it is a commodity.
I kept the money part because, because the goal is that Money be backed by bitcoin and not gold or US/Euro.