contributors: Fix broken test
GitHub recently made a change on their end to the URLs with which
profile images are stored. As a result, this has caused the profile
images on the About Us page to break, and tests to fail.

This updates the URL to what GitHub has switched to, in order to resolve
the issue.