webmaster_tools: Update verification link
This updates an old link to webmaster tools with a new one, so that it
will be possible to tell if there are any crawl errors or 404s on the
site. This is important because if there are, Bitcoin.org can get
penalized in search results thereby making it rank further down below
other Bitcoin websites that are competitively trying to secure top
positions and rank higher.