Change Coinomi wallet link to solve issue #1622
As the policy of Bitcoin.org does not allow closed source
wallets that allow user private key access, we are changing
the link for the Coinomi Android wallet to point to the
Github release page, that contains the latest open source
version along with the necessary source code.

Additionally we revert the transparency to "checkpasstransparencyopensource"
as the source code is available from the same site.