docs: Add FAQ for Translators
This adds a section for frequently asked questions to in the getting
started guide for 'Assisting with Translations' in the GitHub
repository.

Huge thanks goes to @Komodorpudel for leading the effort to put this
together to help make things easier for new contributors to the
translation project.

This will be merged once tests pass.