Tests: fix wallet schema test
The test was designed to run independently and return false on any
failures.  When it was modified to run as a child process of `find`,
the fact that `find` returns true on finding any matching files
prevented any test failures from surfacing.

This restores the test to returing false on any failures by failing
anytime the test prints an error message.  This was tested on the case
that slipped by the test in #2840