community: Remove Bitcoin Organization of the Philippines
This removes the Bitcoin Organization of the Philippines from the
Community page and will be merged once tests pass. Their site has been
going up and down over the past month, and has now been down for seven
days straight, it seems.

This is to protect our page rank from being harmed by linking to web
pages (like this one) that 404 and/or domains that do not respond. When
a page on the site does this, search engines can in turn give our
corresponding linking page a lower search engine ranking, because they
will deem that the page has less relevancy for the key phrase, since it
links to content that doesn't exist (compared to other web pages on
other web sites that could be competing in the search results against
us, that do not).