Add Bitrawr to list of exchange resources
Hey Bitcoin.org!

For the past 6 months, I've been building Bitrawr which acts as a simple knowledge portal for new users. The target audience is for new users to familiarize themselves with the best exchanges, most-trusted wallets, nearest ATMs, efficient mining equipment, beginner guides, and more. This project is continually worked on by me to make sure that it provides the most relevant information for each country.

The site is built on Ruby on Rails. SSL secured.
I would love to hear feedback if any, and also contribute to the Bitcoin space in any way I can.