Update full-node.md
PR: Add a Bitcoin Knots installation tutorial video by btcsessions below the coinkite video tutorial on installing bitcoin core.

Logic for this change: 1) giving representation to knots, which is reflected as a complementary option to core in the bitcoin.org wallet walk through. 2) significant interest in Knots by the general bitcoin community right now, with 15%-21% of the nodes currently.